accessible/tests/mochitest/test_descr.html

<html>

<head>
  <title>nsIAccessible::description tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="common.js"></script>

  <script type="application/javascript">
    function testDescr(aAccOrElmOrID, aDescr)
    {
      var acc = getAccessible(aAccOrElmOrID);
      if (!acc)
        return;

      is(acc.description, aDescr,
         "Wrong description for " + prettyName(aAccOrElmOrID));
    }

    function doTest()
    {
      // Description from aria-describedby attribute
      testDescr("img1", "aria description");

      // No description from @title attribute because it is used to generate
      // name.
      testDescr("img2", "");

      // Description from @title attribute, name is generated from @alt
      // attribute.
      testDescr("img3", "description");

      // Description from content of h2.
      testDescr("p", "heading");

      // From table summary (caption is used as a name)
      testDescr("table1", "summary");

      // Empty (summary is used as a name)
      testDescr("table2", "");

      // From title (summary is used as a name)
      testDescr("table3", "title");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=489944"
     title="@title attribute no longer exposed on accDescription">
    Mozilla Bug 489944
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=666212"
     title="summary attribute content mapped to accessible name in MSAA">
    Mozilla Bug 666212
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <p id="description">aria description</p>
  <img id="img1" aria-describedby="description" />
  <img id="img2" title="title" />
  <img id="img3" alt="name" title="description" />

  <h2 id="heading">heading</h2>
  <p id="p" aria-describedby="heading" role="button">click me</p>

  <table id="table1" summary="summary">
    <caption>caption</caption>
    <tr><td>cell</td></tr>
  </table>

  <table id="table2" summary="summary">
    <tr><td>cell</td></tr>
  </table>

  <table id="table3" summary="summary" title="title">
    <tr><td>cell</td></tr>
  </table>
</body>
</html>
